1.001

+++ src/main/webapp/jsp/listUser.jsp
@@ -8,18 +8,20 @@
 <title>Show All Users</title>
 </head>
 <body>
-<div   class = "pageBody" align = "center">
-           <c:forEach items="${users}" var="user">
-                <div style = "width:70%; height:100;  border: 1px solid black;" >
-                    <img src = "${user.imageURL}" onerror="this.src = '/css/img/default-user-img.jpg'"  height="98" style = "float: left;"/>
+<div class="item2"  style = "background-color: white; ">	</div>
+<div class="item4"></div>
+                              <div class="item3">
+                <c:forEach items="${users}" var="user">
+                <div class = "pageContentUser" >
+                    <img src = "${user.imageURL}" onerror="this.src = '/css/img/default-user-img.jpg'"  class = "imgListUser"/>
                     <h hidden="false" >${user.id}</h>
-                    <h3 style = "margin-top: 0.3; margin-bottom: 0.3;"><a  href = "/showUser?userId=${user.id}">${user.login}</a></h3>
-                    <h style = "margin-top: 0.3; margin-bottom: 0.3;"><fmt:message key = "key.mail" bundle = "${lang}"/>: ${user.mail}</h>
-                    <h style = "margin-top: 0.3; margin-bottom: 0.3;"><fmt:message key = "key.birthDate" bundle = "${lang}"/>: <custom:formatDate value = "${user.birthDate}" pattern = "${dateFormat}"/></h>
-                    <h style = "margin-top: 0.3; margin-bottom: 0.3;"><fmt:message key = "key.registrationDate" bundle = "${lang}"/>: <custom:formatDate value = "${user.registrationDate}" pattern = "${dateFormat}"/></h>
+                    <h3><a  href = "/showUser?userId=${user.id}">${user.login}</a></h3>
+                    <h4><fmt:message key = "key.mail" bundle = "${lang}"/>: ${user.mail}</h4>
+                   <h4><fmt:message key = "key.birthDate" bundle = "${lang}"/>: <custom:formatDate value = "${user.birthDate}" pattern = "${dateFormat}"/></h4>
+                    <h4><fmt:message key = "key.registrationDate" bundle = "${lang}"/>: <custom:formatDate value = "${user.registrationDate}" pattern = "${dateFormat}"/></h4>
                     <p></p>
                     <a href="/showUserEdit?userId=${user.id}"><fmt:message key = "key.edit" bundle = "${lang}"/></a>
-                    <a href="/deleteUser?userId=${user.id}"><fmt:message key = "key.delete" bundle = "${lang}"/></a>
+                    <a  onclick = "return confirm('<fmt:message key = "key.doYouWantReplaceUser" bundle = "${lang}"/>');" href="/deleteUser?userId=${user.id}"><fmt:message key = "key.delete" bundle = "${lang}"/></a>
                 </div>
             </c:forEach>
            </div>
